scripts/requirements.txt
openai==1.43.0
click
python-dotenv

